// src/app/models/aval.ts
export class Aval {
  id: string;
  ofertaId: string;
  //ofertaDescripcion:string;
  fecha: Date;
  fechaSolicitud: Date;
  fechaRecepcion: Date;
  fechaAnulacion: Date;
  fechaSolicitudDevolucion: Date;
  importe: number;
  tipoAvalId: number;
  tipoAval: string
  estadoAvalId: number;
  estadoAval: string;
  avalistaId: number;
  avalista: string
  proyectoId: string;
  proyectoDescripcion:string;
  observaciones: string;
  responsableId:number;
  responsableNombre:string
  clienteId:string;
  clienteNombre:string;
}